retrieval-api/requirements.txt
requests
fastapi
uvicorn
faiss-cpu
numpy
pandas
pydantic
pyarrow